Move halos to shared
